+++ tests/e2e/commands.test.ts
@@ -966,5 +966,26 @@ describe('E2E', () => {
       const result = getCommandOutput(args, { testPath });
       await expect(cleanupOutput(result)).toMatchFileSnapshot(join(testPath, 'snapshot.txt'));
     });
+
+    test('stats should support AsyncAPI 2.x (stylish format)', async () => {
+      const testPath = join(folderPath, 'stats-async2-stylish');
+      const args = getParams(indexEntryPoint, ['stats', 'async.yaml']);
+      const result = getCommandOutput(args, { testPath });
+      await expect(cleanupOutput(result)).toMatchFileSnapshot(join(testPath, 'snapshot.txt'));
+    });
+
+    test('stats should support AsyncAPI 2.x (JSON format)', async () => {
+      const testPath = join(folderPath, 'stats-async2-json');
+      const args = getParams(indexEntryPoint, ['stats', 'async.yaml', '--format=json']);
+      const result = getCommandOutput(args, { testPath });
+      await expect(cleanupOutput(result)).toMatchFileSnapshot(join(testPath, 'snapshot.txt'));
+    });
+
+    test('stats should support AsyncAPI 3.x (stylish format)', async () => {
+      const testPath = join(folderPath, 'stats-async3-stylish');
+      const args = getParams(indexEntryPoint, ['stats', 'asyncapi3.yaml']);
+      const result = getCommandOutput(args, { testPath });
+      await expect(cleanupOutput(result)).toMatchFileSnapshot(join(testPath, 'snapshot.txt'));
+    });
   });
 });